fix: address Gemini PR review comments
- hid_injector (high): spawn + handshake the HID helper OUTSIDE the map lock so
  a stuck helper can't block touch/shutdown across devices; re-acquire to insert
  with a race check (kill the loser) to avoid double-spawn.
- QuiescenceDisabler (high): check initWithName: at the class level
  (instancesRespond) BEFORE alloc, so we never leave an allocated-but-
  uninitialized record for ARC to release (UB). Fixed in both swipePath + drag.
- CommandHandler swipe: parse drag coords via NSNumber.doubleValue/CGFloat to
  keep fractional logical-point precision (matches the tap/longPress fix).
- use-device-interaction: guard normalizePoint against a zero-size rect (no NaN
  coords); don't hijack Enter/Backspace when meta/ctrl/alt is held.
- screen_stream: always route capture size through fit_size (even dims) so a
  None max_size with odd native dims can't make screenrecord fail.
- ActionExecutor: coerce (durationMs/5) as Long before toInt to avoid overflow.

+++ ios-agent/TapsmithAgent/CommandHandler.swift
@@ -791,13 +791,13 @@ class CommandHandler {
         // ─── Swipe / Scroll ───
 
         case "swipe":
-            if let fromX = (params["fromX"] as? NSNumber)?.intValue,
-               let fromY = (params["fromY"] as? NSNumber)?.intValue,
-               let toX = (params["toX"] as? NSNumber)?.intValue,
-               let toY = (params["toY"] as? NSNumber)?.intValue {
+            if let fromX = (params["fromX"] as? NSNumber)?.doubleValue,
+               let fromY = (params["fromY"] as? NSNumber)?.doubleValue,
+               let toX = (params["toX"] as? NSNumber)?.doubleValue,
+               let toY = (params["toY"] as? NSNumber)?.doubleValue {
                 try actionExecutor.drag(
-                    from: CGPoint(x: fromX, y: fromY),
-                    to: CGPoint(x: toX, y: toY)
+                    from: CGPoint(x: CGFloat(fromX), y: CGFloat(fromY)),
+                    to: CGPoint(x: CGFloat(toX), y: CGFloat(toY))
                 )
                 // Match the settle used by the direction-based swipe below.
                 Thread.sleep(forTimeInterval: 0.2)
